+++ CHANGELOG.md
@@ -8,6 +8,8 @@ Check [Keep a Changelog](http://keepachangelog.com/) for recommendations on how 
 ## [Unreleased]
 
 ### Added
+- Skip istio/linkerd-proxy/init container when mirroring a pod without a specific container name.
+- Add "linkerd.io/inject": "disabled" annotation to pod created by mirrord to avoid linkerd auto inject.
 - mirrord-layer: support `-target deployment/deployment_name/container/container_name` flag to run on a specific container.
 - `/nix/*` path is now ignored for file operations to support NixOS.